Restructuring the app to allow creating more modules easily.
