+++ lib/utils.py
@@ -1,0 +1,8 @@
+from termcolor import colored
+
+# Colorize function.
+def color(i, cond):
+  if i != cond:
+    return(colored(i, 'red'))
+  else:
+    return(colored(i, 'green'))
